refactor: break out into seperate DAO vote
